+++ Scripts/Spells/Bushido/SamuraiSpell.cs
@@ -10,7 +10,7 @@ namespace Server.Spells.Bushido
     {
     }
 
-    public abstract double RequiredSkill{ get; }
+    public abstract int RequiredSkill{ get; }
     public abstract int RequiredMana{ get; }
 
     public override SkillName CastSkill => SkillName.Bushido;
